test(libcli): split cli by behaviour family
Split cli.test.js (539 LOC) into cli-parse.test.js (parse + flag/option
scoping + schema validation), cli-help.test.js (showHelp, per-command
help, multi-word) and cli-dispatch.test.js (error/usageError, dispatch,
deps slot, runtime wiring). Shared createProc/definition/createCli
lifted into cli-helpers.js. Assertions unchanged; 27 tests preserved.

Spec 0640 § C (large files), design Decision 5.
